+++ platformio/commands/platform.py
@@ -141,9 +141,8 @@ def platform_list(json_output):
             "url": p.vendor_url,
             # "packages": p.packages.keys(),  # dump all packages
             "packages": p.get_installed_packages().keys(),
-            'forDesktop': any([
-                p.name.startswith(n) for n in ("native", "linux", "windows")
-            ])
+            'forDesktop':
+            any([p.name.startswith(n) for n in ("native", "linux", "windows")])
         })
 
     if json_output: